Support fast path reads for mmapped/borrowed data
Often times an AssetLoader receives a AsyncRead implementation that is
backed by a memory map or a byte slice. In cases where we're dealing
with a format that has a zero-copy parser this introduces an unnecessary
copy. We avoid this by special casing the "vfs" and "dvdbnd" asset
sources to produce a special `FastPathAssetReader`.